feat: support disabling agent retries
we haven't figured a reasonable way to specify model fallbacks in the
parameters. the case in question is one where it runs 
"openrouter:meta-llama/llama-3.3-70b-instruct" once, then falls back
to "google:gemini-2.5-flash". instead of trying openrouter five times,
the fallback should be immediate, but representing this in the API
is harder, and i'd like to get some dependencies off of other SDKs
and just using this library for everythin.